+++ tests/snapshots/snap_test_templates.py
@@ -334,6 +334,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                     </div>
                 </div>
 
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
+                    </div>
+                </div>
+
                 
                 <div class="control-group" data-bind="using: rules.STARTUP">
                     <label class="control-label">on Startup</label>
@@ -737,6 +775,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                             <!--/ko-->
                             
                         </div>
+                    </div>
+                </div>
+
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
                     </div>
                 </div>
 
@@ -1146,6 +1222,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                     </div>
                 </div>
 
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
+                    </div>
+                </div>
+
                 
                 <div class="control-group" data-bind="using: rules.STARTUP">
                     <label class="control-label">on Startup</label>
@@ -1549,6 +1663,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                             <!--/ko-->
                             
                         </div>
+                    </div>
+                </div>
+
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
                     </div>
                 </div>
 
@@ -1958,6 +2110,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                     </div>
                 </div>
 
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
+                    </div>
+                </div>
+
                 
                 <div class="control-group" data-bind="using: rules.STARTUP">
                     <label class="control-label">on Startup</label>
@@ -2361,6 +2551,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                             <!--/ko-->
                             
                         </div>
+                    </div>
+                </div>
+
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
                     </div>
                 </div>
 
@@ -2770,6 +2998,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                     </div>
                 </div>
 
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
+                    </div>
+                </div>
+
                 
                 <div class="control-group" data-bind="using: rules.STARTUP">
                     <label class="control-label">on Startup</label>
@@ -3173,6 +3439,44 @@ snapshots['TestTemplates::test_templates octorelay_settings.jinja2'] = '''<form 
                             <!--/ko-->
                             
                         </div>
+                    </div>
+                </div>
+
+                <div class="control-group">
+                    <label class="control-label">Alert on switches ahead</label>
+                    <div class="controls">
+                        <div class="btn-group">
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-info': show_upcoming() === true,
+                                'btn-default': show_upcoming() !== true
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: true, checked: show_upcoming"
+                                />
+                                YES
+                            </label>
+                            <!--/ko-->
+                            
+                            <!--ko let: { classBinding: {
+                                'active btn-default': show_upcoming() === false,
+                                'btn-default': show_upcoming() !== false
+                            } } -->
+                            <label class="btn" data-bind="css: classBinding">
+                                <input
+                                    type="radio"
+                                    data-bind="checkedValue: false, checked: show_upcoming"
+                                />
+                                NO
+                            </label>
+                            <!--/ko-->
+                            
+                        </div>
+                        <span class="help-inline">
+                            Including the ability to cancel
+                        </span>
                     </div>
                 </div>