+++ ts/OctoPrint.d.ts
@@ -3,6 +3,28 @@
 interface JQuery {
   /** @see https://github.com/OctoPrint/OctoPrint/blob/1.9.0/src/octoprint/static/js/lib/bootstrap/bootstrap-modal.js */
   modal(option: "show" | "hide"): JQuery;
+  tooltip(
+    option:
+      | "destroy"
+      | "toggle"
+      | "hide"
+      | "show"
+      | { placement?: "top" | "bottom" | "left" | "right" }
+  );
+  popover(
+    option:
+      | "destroy"
+      | "toggle"
+      | "hide"
+      | "show"
+      | {
+          html?: boolean;
+          placement?: "top" | "bottom" | "left" | "right";
+          trigger?: "click" | "hover" | "focus" | "manual";
+          title?: string;
+          content?: string;
+        }
+  );
 }
 
 interface ViewModel {
@@ -11,6 +33,7 @@ interface ViewModel {
 }
 
 declare const OCTOPRINT_VIEWMODELS: Array<ViewModel>;
+declare const LOCALE: string;
 
 /** @see https://github.com/OctoPrint/OctoPrint/blob/1.9.0/docs/jsclientlib/base.rst */
 declare const OctoPrint: {